+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning 2.0.0](http://semver.org/).
 
+##	[0.5.0] - Mar 6th, 2018, RISKY
+
+*	README greatly re-constructed.
+*	`(new swift.Connection).createObject()` accepts customised meta data.
+*	`(new swift.Connection).createObject()` validates arguments and may throw error with message "invalid arguments".
+
 ##	[0.4.2] - Feb 8th, 2018
 
 *	Remove redundant debug code which was wrongly published in version 0.4.0 and 0.4.1 .